% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sub_grids.R
\name{gen_all_sub_grids}
\alias{gen_all_sub_grids}
\title{Generate sub-grids for 1x1, 5x5, 10x10, 50x50 and Oklahoma spatial scales}
\usage{
gen_all_sub_grids(dssat_dir = "/carl-data/shared/carl/DSSAT47",
  file_path = c("/Weather/idw.nc", "/Soil/statsgo.nc"),
  filex_path = c("/gridded_DSSAT/grid_filex.nc"), ref_lat = 34.5,
  ref_lon = -99.35, grid_size = c(1, 5, 10, 25, 50))
}
\arguments{
\item{ref_lat}{the latitude in decimal degrees for the lower left point of the sub-grid}

\item{ref_lon}{the longitude in decimal degrees for the lower left point of the sub-grid}

\item{file_name}{the name of the file from which to extract the sub-grid}

\item{no_pts}{the number of points in one side of the sub-grid (no_pts x no_pts square)}
}
\description{
Generate sub-grids for 1x1, 5x5, 10x10, 50x50 and Oklahoma spatial scales
}
