% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sub_grids.R
\name{sub_grid}
\alias{sub_grid}
\title{Create sub-grids of varying spatial scale}
\usage{
sub_grid(file_name, ref_lat, ref_lon, no_pts)
}
\arguments{
\item{file_name}{the name of the file from which to extract the sub-grid}

\item{ref_lat}{the latitude in decimal degrees for the lower left point of the sub-grid}

\item{ref_lon}{the longitude in decimal degrees for the lower left point of the sub-grid}

\item{no_pts}{the number of points in one side of the sub-grid (no_pts x no_pts square)}
}
\description{
Create sub-grids of varying spatial scale
}
